refactor: move the Tor exports into crate::zcash_jni
Move the seventeen `TorClient` and `TorWalletClient` exports out of lib.rs,
together with the marshalling that serves only them: `JNI_HTTP_HEADER`,
`encode_http_header`, `encode_http_response_bytes`,
`encode_address_check_result`, and `parse_http_headers`. lib.rs drops from
3461 to 2711 lines and from 67 exports to 50.

Pure move again: no function body changes, only the new file's docstring and
the imports on both sides.

`parse_tor_dormant_mode` stays behind. It is a `u32` to `DormantMode` match
with no JNI in its signature, so by the boundary rule it is logic, not glue.

Both Java classes land in one submodule because they share that
marshalling and the `TorRuntime` handle; the split is by area, not one file
per class.

Six imports that only the Tor exports used are dropped from lib.rs:
`std::io`, `SystemTime`, `bytes::Bytes`, `http_body_util::BodyExt`,
`ToplevelBlockOn`, and the `tor::http` group.

Exported symbol set unchanged: 78 before, 78 after, empty `nm -gU` diff.

    ./scripts/check-jni-symbols.sh          # 78 symbols, baseline matches
    ./scripts/check-jni-move.sh d65b4d2c    # 78 byte-identical, 0 changed
